assignment-main (1)/assignment-main/App/requirements.txt
langchain
langchain_experimental
langchain_community
streamlit
python-dotenv
langchain_openai
faiss-cpu